+++ extensions/jwt/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.0] - 2025-07-15
+
+- Added support for static headers returned with 401 responses.
+- Renamed the `[extensions.jwt.config.oauth]` section of the configuration to `[extensions.jwt.config.oauth.protected_resource.metadata]`.
+
 ## [1.2.0] - 2025-06-25
 
 - Added support for [OAuth 2.0 Protected Resource Metadata](https://datatracker.ietf.org/doc/html/rfc9728) and returning the `WWW-Authenticate` header on unsuccessful authentication. See the README for the new configuration options.